requirements.txt
google-cloud-language==1.2.0
numpy==1.15.4
praw==6.1.1
